requirements.txt
Django==5.0.1
mailersend==2.0.0
python-dotenv==1.1.1
Pillow==10.1.0